Replace create_path with safe alternative.
create_path() returns a pointer to a local buffer. The OsString it
creates doesn't outlive the pointer to it that is returned; the data can
be overwritten with garbage.

Instead, a new function pathbuf_into_cstring() consumes a PathBuf and
returns a CString, which callers then keep around to keep the memory
safe while using it to get a pointer.