Make tests "pass" on NetBSD
For the nobody uid it's the same as OpenBSD. For the gid, however, it's
a bit different. You can't guarantee the gid is going to be the same
"max value" as per the uid for the nobody user. On my two amd64 NetBSD 8
systems the nobody user is listed as:

nobody:*:32767:39:Unprivileged user:/nonexistent:/sbin/nologin

But I don't _think_ that 39 is going to be guaranteed everywhere.
Therefore simplest solution seems to be to skip this test on NetBSD.